Changing amount to int, rather than double.

+++ src/SFA.DAS.LevyTransferMatching.Database/Tables/Pledge.sql
@@ -2,7 +2,7 @@
     [PledgeId]          INT           IDENTITY (1, 1) NOT NULL,
     [EncodedId]         VARCHAR (100) NOT NULL,
     [EmployerAccountId] BIGINT        NOT NULL,
-    [Amount]            DECIMAL (18)  NOT NULL,
+    [Amount]            INT           NOT NULL,
     [IsNamePublic]      BIT           NOT NULL,
     [CreationDate]      DATETIME2 (7) CONSTRAINT [DF_Pledge__CreationDate] DEFAULT (getdate()) NOT NULL,
     CONSTRAINT [PK_Pledge] PRIMARY KEY CLUSTERED ([PledgeId] ASC)
